Refactor hosted response classification
