+++ packages/client/src/connection-transport.ts
@@ -74,6 +74,7 @@ import {
   retryExplicitConnectorBusy
 } from "./transient-retry.js";
 import { probeLoopbackAccess, tokenSupportsDirectAccess } from "./direct-access.js";
+import { decodeHostedOperationResponse } from "./hosted-operation-response.js";
 
 export type {
   ConnectionTransportInternals,
@@ -324,33 +325,17 @@ export class ConnectionTransport {
         knownRejected
       )
     });
-    let body: any;
-    let malformedHostedHttpFailure = false;
+    let decoded;
     try {
-      body = await decodeJsonResponse(
-        response,
-        "invalid_operation_response",
-        "The collection authority returned a response that is not valid JSON."
-      );
+      decoded = await decodeHostedOperationResponse(response, token.authority !== undefined);
     } catch (cause) {
-      if (!response.ok && token.authority && isDefinitiveHostedHttpRejection(response.status)) {
-        malformedHostedHttpFailure = true;
-      } else {
-        if (attempt.pendingMutation) throw unknownMutationOutcome(attempt.requestId, cause);
-        throw connectError(
-          "invalid_operation_response",
-          "The collection authority returned a response that is not valid JSON.",
-          { cause }
-        );
-      }
+      if (attempt.pendingMutation) throw unknownMutationOutcome(attempt.requestId, cause);
+      throw cause;
     }
+    const body = decoded.body;
     if (!response.ok) {
-      const error = malformedHostedHttpFailure
-        ? serverConnectError("operation_failed", "Collection operation failed.", {
-            status: response.status,
-            operationOutcome: "rejected"
-          })
-        : apiError(body, "operation_failed", "Collection operation failed.", response.status);
+      const error = decoded.httpError
+        ?? apiError(body, "operation_failed", "Collection operation failed.", response.status);
       const recovery = await retryBusy(error);
       if (recovery.retried) return recovery.result;
       if (error.code === "fresh_request_required"
@@ -1006,10 +991,4 @@ export class ConnectionTransport {
       void this.keyStore.delete(pending.keyHandle).catch(() => undefined);
     }
   }
-}
-
-function isDefinitiveHostedHttpRejection(status: number): boolean {
-  return status >= 400
-    && status < 500
-    && ![408, 425, 429].includes(status);
 }
